backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.4

# Phase 1: ML & Embeddings
sentence-transformers>=2.3.1
torch>=2.9.0
transformers>=4.40.0

# Vector Search
faiss-cpu==1.7.4
numpy==1.24.3

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

# Phase 3: LLM Integration
google-generativeai==0.3.2
openai==1.6.1
